Feature: filter request by role, add enum Gender type, update JwFilter logic
