+++ .gitignore
@@ -35,5 +35,3 @@ out/
 
 ### VS Code ###
 .vscode/
-
-src/main/resources/application-phat.properties
