--- README.md
@@ -1,5 +1,0 @@
-# cg-vibely-social-service
-
-vibely-social back-end service
-
-# application.properties: changes active profile
