--- src/main/java/com/cg_vibely_social_service/ServletInitializer.java
@@ -1,13 +1,0 @@
-package com.cg_vibely_social_service;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(CgVibelySocialServiceApplication.class);
-	}
-
-}
