--- src/main/java/com/cg_vibely_social_service/repository/RoleRepository.java
@@ -1,9 +1,0 @@
-package com.cg_vibely_social_service.repository;
-
-import com.cg_vibely_social_service.entity.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
-}
