--- src/main/java/com/cg_vibely_social_service/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.cg_vibely_social_service.repository;
-
-import com.cg_vibely_social_service.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
