--- src/main/java/com/cg_vibely_social_service/service/RoleService.java
@@ -1,4 +1,0 @@
-package com.cg_vibely_social_service.service;
-
-public interface RoleService {
-}
