--- src/main/java/com/cg_vibely_social_service/service/UserService.java
@@ -1,8 +1,0 @@
-package com.cg_vibely_social_service.service;
-
-import com.cg_vibely_social_service.entity.User;
-import org.springframework.security.core.userdetails.UserDetailsService;
-
-public interface UserService extends UserDetailsService, GenericService<User> {
-
-}
