+++ src/main/java/com/demo/dto/request/RequestLoginDto.java
@@ -1,0 +1,17 @@
+package com.demo.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Getter
+@Setter
+public class RequestLoginDto {
+    private String email;
+    private String password;
+}
