+++ src/main/java/com/demo/dto/response/ResponseLoginDto.java
@@ -1,0 +1,19 @@
+package com.demo.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Getter
+@Setter
+public class ResponseLoginDto {
+    private String message;
+    private boolean status;
+    private String email;
+    private String accessToken;
+}
