+++ src/main/java/com/demo/repository/RoleRepository.java
@@ -1,0 +1,9 @@
+package com.demo.repository;
+
+import com.demo.entity.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RoleRepository extends JpaRepository<Role, Long> {
+}
