+++ src/main/java/com/demo/service/RoleService.java
@@ -1,0 +1,4 @@
+package com.demo.service;
+
+public interface RoleService {
+}
