+++ src/main/java/com/demo/service/UserService.java
@@ -1,0 +1,8 @@
+package com.demo.service;
+
+import com.demo.entity.User;
+import org.springframework.security.core.userdetails.UserDetailsService;
+
+public interface UserService extends UserDetailsService, GenericService<User> {
+
+}
