--- src/test/java/com/cg_vibely_social_service/CgVibelySocialServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.cg_vibely_social_service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CgVibelySocialServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
